+++ js/get-game-double.js
@@ -1,9 +1,9 @@
 import getElementFromTemplate from './get-element';
 import renderBlock from './render-block';
 import {Answer} from './data/game-data';
-import footerTemplate from './footer';
-import headerTemplate from './header';
-import statsTemplate from './stats';
+import footerTemplate from './footer-template';
+import headerTemplate from './header-template';
+import statsTemplate from './stats-template';
 import repeatGame from './repeat-game';
 import getNextState from './get-next-state';
 import getNextScreen from './get-next-screen';
